Fix for Issue #27
The root problem here was that when I refactored last year to optimize the hide/unhide algorithm to reduce churn (i.e. re-mounting the same component on the same view should not cause setHidden: to be set to YES, then immediately back to NO for scroll-tick events), I thought I could keep the pool reset operation localized to each view/component combination in the tree.

However, like ComponentKit, this wasn't enough. If there was a deep hierarchy where a component was unmounted from a view, but no other component mounted into that view, then we wouldn't hide. There simply wasn't a good place to put this.