Improve state management
We now no longer always return an optional type if you specify a non-optional state when accessing self.state. This is important since I was finding that I was doing lots of self.state! in my components, and I wanted safer nil handling.

This is a pretty complex diff. We define a specialized extension for the default NoState type of TypedComponent. This specialized extension implements initialState()->NoState by returning a shared value.

What this means: When you build a component that has state, you will fail to compile until you add an initialState() function. Since we provide that guarantee, we are able to avoid having an optional self.state type by default.